# Conflicts:
#	biojava-core/src/main/java/org/biojava/nbio/core/sequence/io/GenericInsdcHeaderFormat.java
#	biojava-core/src/test/java/org/biojava/nbio/core/sequence/io/GenbankReaderTest.java

+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/geometry/SuperPosition.java
@@ -145,7 +145,7 @@ public final class SuperPosition {
         for (int i = 0; i < x.length; i++) {
             sum += x[i].distanceSquared(y[i]);
         }
-        return (double)Math.sqrt(sum/x.length);
+        return Math.sqrt(sum/x.length);
     }
 
     public static double rmsdMin(Point3d[] x, Point3d[] y) {
@@ -157,7 +157,7 @@ public final class SuperPosition {
             }
             sum += minDist;
         }
-        return (double)Math.sqrt(sum/x.length);
+        return Math.sqrt(sum/x.length);
     }
 
     /**
